r move API input related code to 'input' module
