+++ src/api/input/request.rs
@@ -1,7 +1,7 @@
 use serde::de::DeserializeOwned;
 use serde_json::{Map, Value};
+use crate::api::input::request_header::RequestHeader;
 
-use crate::api::request_header::RequestHeader;
 use crate::error::{Error, ErrorKind};
 
 #[derive(Debug)]
