+++ src/api/mod.rs
@@ -1,14 +1,4 @@
 pub mod amqp_connection_manager;
-pub mod amqp_request_dispatch;
-pub mod amqp_request_replier;
-pub mod authorizer;
-pub mod element;
 pub mod init;
-pub mod request;
-pub mod request_header;
-pub mod request_result;
-pub mod sanitizer;
-pub mod token;
-pub mod token_validator;
-pub mod request_result_error;
 pub mod initialization_package;
+pub mod input;
